+++ modules/casper.js
@@ -165,7 +165,6 @@ var Casper = function Casper(options) {
         var notices = [];
         if (match && match.length === 4) {
             notices.push('  in module ' + match[2]);
-            notices.push('  NOTICE: errors within modules cannot be backtraced yet.');
             msg = match[3];
         }
         console.error(c.colorize(msg, 'RED_BAR', 80));